+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2022-08-20
+### Fixed
+- Fix an incorrect zooming behaviour with Canvas Scalar.
+### Added
+- An Unity Event API for scale callback.
+- Set Scale API Method for external control for zooming.
+
 ## [1.0.2] - 2022-07-16
 ### Fixed
 - Fix an incorrect zooming behaviour when Content's RectTransform has a stretched anchor.